pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "germannouns"
version = "0.1.0"
description = "German noun gender classification using machine learning"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Sozair Ali", email = "sozairali@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "pandas>=1.5.0",
    "scikit-learn>=1.2.0",
    "joblib>=1.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
data = [
    "openai>=1.0.0",
    "tensorflow>=2.10.0",
]
web = [
    "flask>=2.3.0",
    "gunicorn>=20.1.0",
]
dev = [
    "pytest>=7.3.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
all = [
    "germannouns[data,web,dev]",
]

[project.scripts]
germannouns-train = "germannouns.model.trainer:main"
germannouns-predict = "germannouns.model.predictor:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
